// src/app/invoices/models/invoicesresponse.ts
import { ISystemActivities } from '../../shared/models/systemactivitiesrequest';
import { IFeeDetailsResponse } from "./feedetailsresponse";
import { IPaymentDetailsResponse } from "./paymentdetailsresponse";
import { ITripDetailsResponse } from "./tripdetailsresponse";

export interface IInvoiceResponse {
    InvoiceId: number;
    InvoiceNumber: string;
    CustomerId: number;
    InvoiceOption: string;
    StartDate: Date;
    EndDate: Date;
    DueDate: Date;
    InvoiceDate: Date;
    CreatedDate: Date;
    Status: string;
    TotalCount: number;
    PlateNumber: string;
    CitationId: number;
    VehicleId: number;
    CustomerStatus: string;
    IsCustomerInvoice: boolean;
    InvoiceStatus: string;
    Isviolator: boolean;
    InvBatchId: string;
    Stmt_Literal: string;
    SearchActivityIndicator: boolean;
    PageSize: number;
    PageNumber: number;
    SortColumn: string;
    SortDirection: number;
    InvOutstading: number;
    InvStatus: string;
    InvoicePaymentAmount: number;
    BalanceDue: number;
    AmountPaid: number;
    TotalAmount: number;
    SystemActivity: ISystemActivities;
    UserName: string;
    DocumentPath: string;
    FirstName: string;
    LastName: string;
    objFee: IFeeDetailsResponse[];
    objPayments: IPaymentDetailsResponse[];
    objTrips: ITripDetailsResponse[];
    AdjustmentAmount: number;
    BilledAmount: number;
    PreviousDue: number;
    IsChecked: boolean;
    OutstandingDue: number;
    AdjTotalAmount: number;
    InvoiceBatchId: string;
    AgingHoldType: string;
    HoldStatus: string;
    IsAgingHold: boolean;
    TotalPmtAdjAmt: number;
    UnbilledAmount: number;
    ProblemID: number;
    InvoiceChecked: boolean;
    ActualDueDate: Date;
    MaxExtensionDate: Date;
    GracePeriodDays: number;
}